+++ src/pages/viewer/Channels.vue
@@ -46,16 +46,18 @@
           </q-popover>
         </q-item>
         <transition appear enter-active-class="animated bounceInDown" leave-active-class="animated bounceOutUp" v-if="$q.platform.is.desktop">
-          <div title="View hex payload" class="on-right cursor-pointer pull-right text-center round-borders q-px-xs" v-if="selectedItem && selectedItem.protocol_id === proxyProtocolId" @click="hexViewHandler">
+          <q-btn title="View hex payload" class="on-right pull-right text-center round-borders q-px-xs q-py-none" v-if="selectedItem && selectedItem.protocol_id === proxyProtocolId" @click="hexViewHandler" flat dense style="width: 50px;">
             <q-icon size="1.5rem" color="white" name="mdi-matrix"/>
-            <div style="font-size: .9rem;">HEX</div>
-          </div>
+            <div style="font-size: .7rem;">HEX</div>
+          </q-btn>
         </transition>
       </div>
       <div v-if="active">
-        <q-btn title="Mode (Real-time/History)" size="sm" v-if="!selectedItem.deleted" flat class="on-left" color="white" @click="modeModel = !modeModel" :icon="modeModel ? 'playlist_play' : 'history'" :rounded="$q.platform.is.mobile">
-          {{$q.platform.is.mobile ? '' : modeModel ? 'Real-time' : 'History'}}
-          <q-chip small square color="red" v-if="newMessagesCount" class="cursor-pointer q-ml-sm">{{newMessagesCount}}</q-chip>
+        <q-btn v-if="!selectedItem.deleted" flat dense class="on-right pull-right text-center round-borders q-px-xs q-py-none" color="white" @click="modeModel = !modeModel" style="min-width: 70px; max-width: 70px;">
+          <q-icon size="1.5rem" color="white" :name="modeModel ? 'playlist_play' : 'history'"/>
+          <div style="font-size: .7rem;">{{modeModel ? 'Real-time' : 'History'}}</div>
+          <div class="bg-red text-white q-pa-xs round-borders cursor-pointer absolute-top-right" v-if="newMessagesCount" style="font-size: .6rem;">{{newMessagesCount}}</div>
+          <q-tooltip>Mode (Real-time/History)</q-tooltip>
         </q-btn>
         <q-btn-toggle
           v-if="!selectedItem.deleted"
@@ -86,10 +88,10 @@
       </div>
       <div v-if="active && $q.platform.is.desktop" class="flex" style="width: 46px;">
         <transition appear enter-active-class="animated bounceInDown" leave-active-class="animated bounceOutUp">
-          <div title="Clear all panes" class="on-left cursor-pointer pull-right text-center" v-if="modeModel && !isEmptyMessages" @click="clearHandler">
+          <q-btn title="Clear all panes" class="on-left pull-right text-center q-py-none" v-if="modeModel && !isEmptyMessages" @click="clearHandler" flat dense style="width: 60px">
             <q-icon size="1.5rem" color="white" name="mdi-playlist-remove"/>
-            <div style="font-size: .9rem;">Clear</div>
-          </div>
+            <div style="font-size: .7rem;">Clear</div>
+          </q-btn>
         </transition>
       </div>
       <div v-else-if="active && !$q.platform.is.desktop">
@@ -126,6 +128,7 @@
         ref="messages"
         @view-data="viewDataHandler"
         :mode="mode"
+        :item="selectedItem"
         :activeId="active"
         :isEnabled="!!+size[1]"
         :limit="limit"